great success! visible, although a bit warped Mandelbrot set fractal there; added an inconsequential (long double) cast to scaler

+++ srcs/scale.c
@@ -6,11 +6,11 @@
 /*   By: thakala <thakala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 14:15:58 by thakala           #+#    #+#             */
-/*   Updated: 2022/09/29 14:25:21 by thakala          ###   ########.fr       */
+/*   Updated: 2022/09/29 17:23:48 by thakala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 long double	scale(long double low, int interval, int pos, long double high)
 {
-	return ((high - low) / (long double) interval * pos);
+	return ((high - low) / (long double) interval * (long double) pos);
 }
